feat: add Admin auth category CRUD

+++ src/categories/categories.controller.ts
@@ -6,6 +6,9 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import {
   ApiTags,
@@ -13,16 +16,27 @@ import {
   ApiResponse,
   ApiBody,
   ApiParam,
+  ApiBearerAuth,
 } from '@nestjs/swagger';
 import { CategoriesService } from './categories.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { ClsService } from 'nestjs-cls';
+import { AppClsStore, UserType } from 'src/Types/user.types';
+import { UserService } from 'src/user/user.service';
 
 @ApiTags('categories')
 @Controller({ path: 'categories', version: '1' })
 export class CategoriesController {
-  constructor(private readonly categoriesService: CategoriesService) {}
+  constructor(
+    private readonly categoriesService: CategoriesService,
+    private readonly clsService: ClsService,
+    private readonly userService: UserService,
+  ) {}
 
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard('jwt'))
   @Post()
   @ApiOperation({ summary: 'Create a new category' })
   @ApiResponse({
@@ -31,7 +45,19 @@ export class CategoriesController {
   })
   @ApiResponse({ status: 400, description: 'Bad Request.' })
   @ApiBody({ type: CreateCategoryDto })
-  create(@Body() createCategoryDto: CreateCategoryDto) {
+  async create(@Body() createCategoryDto: CreateCategoryDto) {
+    const context = this.clsService.get<AppClsStore>();
+    if (!context || !context.user) {
+      throw new HttpException('User not found', HttpStatus.BAD_REQUEST);
+    }
+    const user = await this.userService.findOne({ userId: context.user.id });
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    if (user.type != UserType.Admin) {
+      throw new HttpException('UNAUTHORIZED User', HttpStatus.UNAUTHORIZED);
+    }
+
     return this.categoriesService.create(createCategoryDto);
   }
 
@@ -54,25 +80,51 @@ export class CategoriesController {
     return this.categoriesService.findOne(id);
   }
 
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard('jwt'))
   @Patch(':id')
   @ApiOperation({ summary: 'Update a category by id' })
   @ApiResponse({ status: 200, description: 'Category updated successfully.' })
   @ApiResponse({ status: 404, description: 'Category not found.' })
   @ApiParam({ name: 'id', description: 'The id of the category to update' })
   @ApiBody({ type: UpdateCategoryDto })
-  update(
+  async update(
     @Param('id') id: string,
     @Body() updateCategoryDto: UpdateCategoryDto,
   ) {
+    const context = this.clsService.get<AppClsStore>();
+    if (!context || !context.user) {
+      throw new HttpException('User not found', HttpStatus.BAD_REQUEST);
+    }
+    const user = await this.userService.findOne({ userId: context.user.id });
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    if (user.type != UserType.Admin) {
+      throw new HttpException('UNAUTHORIZED User', HttpStatus.UNAUTHORIZED);
+    }
     return this.categoriesService.update(id, updateCategoryDto);
   }
 
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard('jwt'))
   @Delete(':id')
   @ApiOperation({ summary: 'Delete a category by id' })
   @ApiResponse({ status: 200, description: 'Category deleted successfully.' })
   @ApiResponse({ status: 404, description: 'Category not found.' })
   @ApiParam({ name: 'id', description: 'The id of the category to delete' })
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const context = this.clsService.get<AppClsStore>();
+    if (!context || !context.user) {
+      throw new HttpException('User not found', HttpStatus.BAD_REQUEST);
+    }
+    const user = await this.userService.findOne({ userId: context.user.id });
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    if (user.type != UserType.Admin) {
+      throw new HttpException('UNAUTHORIZED User', HttpStatus.UNAUTHORIZED);
+    }
     return this.categoriesService.remove(id);
   }
 }
